# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a Bible API system, powered by Gin and GORM.",
        "title": "Bible System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/your-repo",
            "email": "rayselfs@alive.org.tw"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/bible/v1/search": {
            "post": {
                "description": "Use semantic search functionality to understand query meaning and return relevant verses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bible"
                ],
                "summary": "Semantic search for Bible verses",
                "parameters": [
                    {
                        "description": "Search request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "schema": {
                            "$ref": "#/definitions/models.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Search failed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/bible/v1/version/{version_id}": {
            "get": {
                "description": "Get all books, chapters and verses content for the specified version ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bible"
                ],
                "summary": "Get complete Bible content",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Version ID",
                        "name": "version_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved Bible content",
                        "schema": {
                            "$ref": "#/definitions/models.BibleContentAPI"
                        }
                    },
                    "400": {
                        "description": "Invalid input parameters",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/bible/v1/versions": {
            "get": {
                "description": "List all available Bible versions in the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bible"
                ],
                "summary": "Get all Bible versions",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved version list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.VersionListItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BibleContentAPI": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BibleContentBook"
                    }
                },
                "version_code": {
                    "type": "string"
                },
                "version_id": {
                    "type": "integer"
                },
                "version_name": {
                    "type": "string"
                }
            }
        },
        "models.BibleContentBook": {
            "type": "object",
            "properties": {
                "abbreviation": {
                    "type": "string"
                },
                "chapters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BibleContentChapter"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                }
            }
        },
        "models.BibleContentChapter": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "verses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BibleContentVerse"
                    }
                }
            }
        },
        "models.BibleContentVerse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.SearchRequest": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "query": {
                    "type": "string"
                },
                "top_k": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.SearchResponse": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SearchResult"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.SearchResult": {
            "type": "object",
            "properties": {
                "book": {
                    "type": "integer"
                },
                "chapter": {
                    "type": "integer"
                },
                "rank": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "text": {
                    "type": "string"
                },
                "verse": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.VersionListItem": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        }
    }
}